refactor: allow adding openai compatible providers in onboarding, fix DRY violations
